Chart aggregate and discrete session usage
Currently, CurrentSessions is completely ignored. This makes it hard to
measure overall utilization.

This PR adds two new metrics:

passenger.process.sessions - the number of sessions for a discrete
process

passenger.sessions.total - the sum of sessions for the entire host

I'm not sure both are strictly necessary, but I am trying to match the
existing metrics as closely as possible.